feature: creating official home page
